# backend/docs/swagger.yaml
basePath: /
definitions:
  controller.LogsResponseWrapper:
    properties:
      code:
        example: 200
        type: integer
      data:
        $ref: '#/definitions/storage.LogQueryResponse'
      message:
        example: success
        type: string
    type: object
  controller.OverviewResponseWrapper:
    properties:
      code:
        example: 200
        type: integer
      data:
        $ref: '#/definitions/storage.OverviewStats'
      message:
        example: success
        type: string
    type: object
  controller.StatsResponseWrapper:
    properties:
      code:
        example: 200
        type: integer
      data:
        additionalProperties: true
        type: object
      message:
        example: success
        type: string
    type: object
  controller.StatsWSInitResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/storage.RealTimeHistoryPoint'
        type: array
      type:
        example: init
        type: string
    type: object
  controller.StatsWSUpdateResponse:
    properties:
      data:
        $ref: '#/definitions/storage.RealTimeHistoryPoint'
      type:
        example: update
        type: string
    type: object
  controller.StatusDistributionResponseWrapper:
    properties:
      code:
        example: 200
        type: integer
      data:
        $ref: '#/definitions/storage.StatusDistributionResult'
      message:
        example: success
        type: string
    type: object
  controller.TimeSeriesResponseWrapper:
    properties:
      code:
        example: 200
        type: integer
      data:
        $ref: '#/definitions/storage.TimeSeriesResult'
      message:
        example: success
        type: string
    type: object
  controller.TopPathsResponseWrapper:
    properties:
      code:
        example: 200
        type: integer
      data:
        items:
          $ref: '#/definitions/storage.TopPathItem'
        type: array
      message:
        example: success
        type: string
    type: object
...
  storage.GeoDistributionItem:
    properties:
      city:
        type: string
      count:
        type: integer
      country:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      region:
        type: string
    type: object
  storage.LogQueryItem:
    properties:
      entry:
        $ref: '#/definitions/model.LogEntry'
      source_id:
        type: string
    type: object
  storage.LogQueryResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/storage.LogQueryItem'
        type: array
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
  storage.MetricValue:
    properties:
      compare_percent:
        type: number
      value:
        type: number
    type: object
  storage.OverviewStats:
    properties:
      avg_response_time:
        $ref: '#/definitions/storage.MetricValue'
      compare_type:
        description: e.g., "vs yesterday" or "vs previous period"
        type: string
      error_rate:
        $ref: '#/definitions/storage.MetricValue'
      success_rate:
        $ref: '#/definitions/storage.MetricValue'
      total_requests:
        $ref: '#/definitions/storage.MetricValue'
    type: object
  storage.RealTimeHistoryPoint:
    properties:
      errors:
        type: integer
      total:
        type: integer
      ts:
        type: string
    type: object
  storage.StatusDistributionEntry:
    properties:
      code_class:
        type: string
      count:
        type: integer
      percentage:
        type: number
    type: object
  storage.StatusDistributionResult:
    properties:
      distribution:
        items:
          $ref: '#/definitions/storage.StatusDistributionEntry'
        type: array
      total:
        type: integer
    type: object
  storage.TimeSeriesPoint:
    properties:
      ts:
        type: string
      value:
        type: number
    type: object
  storage.TimeSeriesResult:
    properties:
      interval:
        type: string
      metric:
        type: string
      points:
        items:
          $ref: '#/definitions/storage.TimeSeriesPoint'
        type: array
    type: object
  storage.TopPathItem:
    properties:
      avg_response_time:
        type: number
      error_rate:
        type: number
      request_count:
        type: integer
      uri:
        type: string
    type: object
...
info:
  contact: {}
  description: SvcWatch is a real-time Nginx log monitoring system.
  title: SvcWatch API
  version: "1.0"
paths:
  /api/v1/sev/distribution:
    get:
      description: Get distribution of status codes for a time range
      parameters:
      - default: "2026-03-19 00:00:00"
        description: Start Time
        in: query
        name: start_time
        required: true
        type: string
      - default: "2026-03-20 00:00:00"
        description: End Time
        in: query
        name: end_time
        required: true
        type: string
      - default: access.log
        description: Log File or Source ID (optional)
        in: query
        name: log_file
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.StatusDistributionResponseWrapper'
      security:
      - BearerAuth: []
      summary: Get HTTP status code distribution
      tags:
      - 2. Overview Dashboard
  /api/v1/sev/logs:
    get:
      description: Query parsed Nginx logs with comprehensive filtering, sorting,
        and pagination
      parameters:
      - default: 1
        description: Page number (default 1)
        in: query
        name: page
        type: integer
      - default: 50
        description: Page size (default 50, max 500)
        in: query
        name: size
        type: integer
      - description: Start Time
        example: "2026-03-19 00:00:00"
        in: query
        name: start_time
        type: string
      - description: End Time
        example: "2026-03-20 00:00:00"
        in: query
        name: end_time
        type: string
      - description: Log File or Source ID
        in: query
        name: source_id
        type: string
      - description: IP address (supports prefix match)
        example: 192.168.1.1
        in: query
        name: ip
        type: string
      - description: HTTP Method (e.g. GET)
        example: GET
        in: query
        name: method
        type: string
      - description: Exact HTTP Status (e.g. 500)
        example: 200
        in: query
        name: status
        type: integer
      - description: HTTP Status Class (e.g. 5xx)
        example: 5xx
        in: query
        name: status_class
        type: string
      - description: Keyword to search in URL path
        example: api
        in: query
        name: path_keyword
        type: string
      - description: Minimum Latency in ms
        example: 100
        in: query
        name: min_latency
        type: integer
      - description: Maximum Latency in ms
        example: 5000
        in: query
        name: max_latency
        type: integer
      - default: time_desc
        description: Sort order (time_desc or latency_desc)
        enum:
        - time_desc
        - latency_desc
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.LogsResponseWrapper'
      security:
      - BearerAuth: []
      summary: Query detailed logs
      tags:
      - 3. Logs Analysis
  /api/v1/sev/logs/ws:
    get:
      description: Upgrade connection to WebSocket and stream raw logs in real-time.
        Pushes standard Nginx log lines as text frames.
      parameters:
      - default: access.log
        description: Log File or Source ID (optional)
        in: query
        name: log_file
        type: string
      responses:
        "101":
          description: Switching Protocols (Handshake Success). Streams raw nginx
            log lines as text frames.
          schema:
            type: string
      summary: Real-time logs streaming via WebSocket
      tags:
      - 2. Overview Dashboard
  /api/v1/sev/overview:
    get:
      description: Get overview statistics with comparison for a time range
      parameters:
      - default: "2026-03-19 00:00:00"
        description: Start Time
        in: query
        name: start_time
        required: true
        type: string
      - default: "2026-03-20 00:00:00"
        description: End Time
        in: query
        name: end_time
        required: true
        type: string
      - default: access.log
        description: Log File or Source ID (optional)
        in: query
        name: log_file
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.OverviewResponseWrapper'
      security:
      - BearerAuth: []
      summary: Get business overview key metrics
      tags:
      - 2. Overview Dashboard
...
  /api/v1/sev/stats:
    get:
      description: Get total logs count for all monitored tables
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.StatsResponseWrapper'
      security:
      - BearerAuth: []
      summary: Get aggregated logs statistics
      tags:
      - 5. System Stats
  /api/v1/sev/stats/geo-distribution:
    get:
      description: Get geographical distribution of IP addresses from logs.
      parameters:
      - default: "2026-03-19 00:00:00"
        description: Start Time
        in: query
        name: start_time
        required: true
        type: string
      - default: "2026-03-20 00:00:00"
        description: End Time
        in: query
        name: end_time
        required: true
        type: string
      - default: access.log
        description: Log File or Source ID
        in: query
        name: source_id
        type: string
      - default: 100
        description: Number of locations to return (default 100, max 1000)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/storage.GeoDistributionItem'
            type: array
      security:
      - BearerAuth: []
      summary: Get geographical distribution of requests
      tags:
      - 4. Geo Analysis
  /api/v1/sev/stats/timeseries:
    get:
      description: Get time-series data for a metric (qps, error_rate, latency_p99,
        bandwidth). Range cannot exceed 1 year. Returns exactly 30 points.
      parameters:
      - description: Metric type
        enum:
        - qps
        - error_rate
        - latency_p99
        - bandwidth
        in: query
        name: metric
        required: true
        type: string
      - default: "2026-03-19 00:00:00"
        description: Start Time
        in: query
        name: start_time
        required: true
        type: string
      - default: "2026-03-20 00:00:00"
        description: End Time
        in: query
        name: end_time
        required: true
        type: string
      - collectionFormat: csv
        description: List of Source IDs or Log Files to aggregate
        in: query
        items:
          type: string
        name: source_ids
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.TimeSeriesResponseWrapper'
      security:
      - BearerAuth: []
      summary: Get trend data for charts
      tags:
      - 2. Overview Dashboard
  /api/v1/sev/stats/top-paths:
    get:
      description: Get the top requested interface URIs along with their request count,
        average response time, and error rate.
      parameters:
      - default: "2026-03-19 00:00:00"
        description: Start Time
        in: query
        name: start_time
        required: true
        type: string
      - default: "2026-03-20 00:00:00"
        description: End Time
        in: query
        name: end_time
        required: true
        type: string
      - default: access.log
        description: Log File or Source ID
        in: query
        name: source_id
        type: string
      - default: 10
        description: Number of top paths to return (default 10, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.TopPathsResponseWrapper'
      security:
      - BearerAuth: []
      summary: Get top requested paths
      tags:
      - 2. Overview Dashboard
  /api/v1/sev/stats/ws:
    get:
      description: Upgrade connection to WebSocket and stream real-time request counts
        and error counts. Pushes initial history as 'init' event and periodic updates
        as 'update' event.
      parameters:
      - default: access.log
        description: Log File or Source ID (optional)
        in: query
        name: log_file
        type: string
      - default: 1s
        description: Refresh interval (e.g., 1s, 2s, 5s)
        in: query
        name: interval
        type: string
      - default: false
        description: Enable mock simulation data
        in: query
        name: simulate
        type: boolean
      responses:
        "101":
          description: Switching Protocols (Handshake Success). Initial history data
            frame pushed immediately.
          schema:
            $ref: '#/definitions/controller.StatsWSInitResponse'
        "200":
          description: Periodic real-time update data frame pushed at the specified
            interval.
          schema:
            $ref: '#/definitions/controller.StatsWSUpdateResponse'
      summary: Real-time statistics push via WebSocket
      tags:
      - 2. Overview Dashboard
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...